+++ composer.json
@@ -12,6 +12,11 @@
     "geoip2/geoip2": "~2.0",
     "jenssegers/agent": "~2.6"
   },
+  "autoload": {
+    "psr-4": {
+      "IgniterLabs\\VisitorTracker\\": "src/"
+    }
+  },
   "extra": {
     "tastyigniter-extension": {
       "code": "igniterlabs.visitortracker",
