+++ resources/views/_partials/dashboardwidgets/pageviews/pageviews.php
