+++ resources/views/pageviews/index.blade.php
