+++ resources/views/pagevisits/index.blade.php
